api/stats/:country working with lower case and titles with spaces
