+++ controllers/helpers/statsHelper.js
@@ -1,10 +1,14 @@
 const axios = require('axios');
+const { capitalize } = require('./helper');
 
 
 module.exports = {
     getDataByLocationAPI: async function (location) {
         //NOTE: This API only searches for the country if it is capitalized.
         //For example, canada should be Canada. 
+       
+        const locCap = await capitalize(location);
+        
         try {
             const apiResponse = await axios({
                 "method": "GET",
@@ -16,7 +20,7 @@ module.exports = {
                     "useQueryString": true
                 },
                 "params": {
-                    "country": location
+                    "country": locCap
                 }
             });
             return apiResponse.data.data;
@@ -24,33 +28,5 @@ module.exports = {
         } catch (error) {
             console.error(error);
         }
-    },
-    splitDBDate: async function (str) {
-        //Takes date string from API, splits it up and returns a new object
-        const [year, month, extra] = str.split("-");
-        const [day] = extra.split("T");
-
-        const dateObj = {
-            "day": parseInt(day),
-            "month": parseInt(month),
-            "year": parseInt(year)
-        }
-
-        return dateObj;
-    },
-    getTodaysDate: async function () {
-        //Gets today's dd/mm/yyyy and returns it as an object
-        const today = new Date();
-        var dd = today.getDate();
-        var mm = today.getMonth() + 1;
-        var yyyy = today.getFullYear();
-
-        const dateObj = {
-            "day": parseInt(dd),
-            "month": parseInt(mm),
-            "year": parseInt(yyyy)
-        }
-
-        return dateObj;
     }
 }
